Adds a basic mechanism for cache-busting images
Seems some old rockylinux image are having problems w corepack and node.
Ensuring the base image is up-to-date seems to fix it. Thus this
feature.

The cache buster actually isn't updated every build, it's just there for
when we ever need to do this globally across all builds again we can go
and update the filename.